+++ src/main/java/net/lopymine/pe/config/ParticleEffectsConfig.java
@@ -43,7 +43,7 @@ public class ParticleEffectsConfig {
 	private static @NotNull ParticleEffectsConfig create() {
 		ParticleEffectsConfig config = new ParticleEffectsConfig();
 		try (FileWriter writer = new FileWriter(CONFIG_FILE, StandardCharsets.UTF_8)) {
-			String json = GSON.toJson(CODEC.encode(config, JsonOps.INSTANCE, JsonOps.INSTANCE.empty())/*? if >=1.20.5 {*/.getOrThrow());/*?} else*//*.getOrThrow(false, LOGGER::error));*/
+			String json = GSON.toJson(CODEC.encode(config, JsonOps.INSTANCE, JsonOps.INSTANCE.empty())/*? if >=1.20.5 {*/.getOrThrow());/*?} else*///.getOrThrow(false, LOGGER::error));
 			writer.write(json);
 		} catch (Exception e) {
 			LOGGER.error("Failed to create config", e);
@@ -68,7 +68,7 @@ public class ParticleEffectsConfig {
 		ParticleEffects.setConfig(this);
 		CompletableFuture.runAsync(() -> {
 			try (FileWriter writer = new FileWriter(CONFIG_FILE, StandardCharsets.UTF_8)) {
-				String json = GSON.toJson(CODEC.encode(this, JsonOps.INSTANCE, JsonOps.INSTANCE.empty())/*? if >=1.20.5 {*/.getOrThrow());/*?} else*//*.getOrThrow(false, LOGGER::error));*/
+				String json = GSON.toJson(CODEC.encode(this, JsonOps.INSTANCE, JsonOps.INSTANCE.empty())/*? if >=1.20.5 {*/.getOrThrow());/*?} else*///.getOrThrow(false, LOGGER::error));
 				writer.write(json);
 			} catch (Exception e) {
 				LOGGER.error("Failed to save config", e);