+++ src/main/java/net/lopymine/pe/entrypoint/PEFabricClientEntrypoint.java
@@ -3,7 +3,9 @@ package net.lopymine.pe.entrypoint;
 //? if fabric {
 
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
 import net.lopymine.pe.client.ParticleEffectsClient;
+import net.lopymine.pe.compat.LoadedMods;
 import net.lopymine.pe.manager.ParticleEffectsManager;
 
 public class PEFabricClientEntrypoint implements ClientModInitializer {
@@ -13,9 +15,21 @@ public class PEFabricClientEntrypoint implements ClientModInitializer {
 		ParticleEffectsClient.onInitializeClient();
 		ParticleEffectsManager.registerParticleTypes();
 		ParticleEffectsManager.swapParticleTypes();
-		ParticleEffectsManager.registerParticleColorsForTypes();
+		if (LoadedMods.isAnyOldPotionsModLoaded()) {
+			ClientLifecycleEvents.CLIENT_STARTED.register((mc) -> {
+				ParticleEffectsManager.redirectEnabled = true;
+				ParticleEffectsManager.redirectToVanillaEffectColors = true;
+				ParticleEffectsManager.registerParticleColorsForTypes();
+				ParticleEffectsManager.redirectToVanillaEffectColors = false;
+				ParticleEffectsManager.registerParticleColorsForTypes();
+				ParticleEffectsManager.redirectEnabled = false;
+			});
+		} else {
+			ParticleEffectsManager.registerParticleColorsForTypes();
+		}
 		ParticleEffectsManager.registerParticleFactories();
 	}
+
 }
 
 //?}